more fixes

+++ app/src/main/java/com/gradysbooch/restaurant/viewmodel/OrderViewModel.kt
@@ -219,23 +219,27 @@ class OrderViewModel(application: Application) : BaseViewModel(application),
     override fun lockOrder(tableUID: String, color: Color)
     {
         viewModelScope.launch {
-            val order = repository.networkRepository
+            if (repository.orderDao().getOrder(tableUID, color) != null) {
+                val order = repository.networkRepository
                     .clientOrders()
                     .first()
                     .find { it.tableUID == tableUID && it.orderColor == color }
                     ?: return@launch
-            repository.networkRepository.lockOrder(tableUID, color)
-            repository.orderDao().addOrder(order)
-            Log.d(this::class.simpleName, "Locked order $tableUID - $color")
+                repository.networkRepository.lockOrder(tableUID, color)
+                repository.orderDao().addOrder(order)
+                Log.d(this::class.simpleName, "Locked order $tableUID - $color")
+            }
         }
     }
 
     override fun unlockOrder(tableUID: String, color: Color)
     {
         viewModelScope.launch {
-            repository.orderDao().deleteOrder(tableUID, color)
-            repository.networkRepository.unlockOrder(tableUID, color)
-            Log.d(this::class.simpleName, "Unlocked order $tableUID - $color")
+            if (repository.orderDao().getOrder(tableUID, color) != null) {
+                repository.orderDao().deleteOrder(tableUID, color)
+                repository.networkRepository.unlockOrder(tableUID, color)
+                Log.d(this::class.simpleName, "Unlocked order $tableUID - $color")
+            }
         }
     }
